fix: Bot.maxCapitalUsd Double -> BigDecimal for Hibernate compatibility

+++ src/main/java/com/trading/assistant/user/model/Bot.java
@@ -1,6 +1,7 @@
 package com.trading.assistant.user.model;
 
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -34,7 +35,7 @@ public class Bot {
     private boolean running = true;
 
     @Column(precision = 20, scale = 8)
-    private Double maxCapitalUsd;
+    private BigDecimal maxCapitalUsd;
 
     @Column(nullable = false, updatable = false)
     private LocalDateTime createdAt;
@@ -78,8 +79,8 @@ public class Bot {
     public boolean isRunning() { return running; }
     public void setRunning(boolean running) { this.running = running; }
 
-    public Double getMaxCapitalUsd() { return maxCapitalUsd; }
-    public void setMaxCapitalUsd(Double maxCapitalUsd) { this.maxCapitalUsd = maxCapitalUsd; }
+    public BigDecimal getMaxCapitalUsd() { return maxCapitalUsd; }
+    public void setMaxCapitalUsd(BigDecimal maxCapitalUsd) { this.maxCapitalUsd = maxCapitalUsd; }
 
     public LocalDateTime getCreatedAt() { return createdAt; }
     public LocalDateTime getUpdatedAt() { return updatedAt; }
